+++ packages/gatsby-source-contentful/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.48](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.47...gatsby-source-contentful@2.0.48) (2019-04-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
 ## [2.0.47](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.46...gatsby-source-contentful@2.0.47) (2019-04-09)
 
 **Note:** Version bump only for package gatsby-source-contentful